[NOM-23] - Moved screen_scaffold to widgets directory
